# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Список",
                "operationId": "76915b004841c2ef5414f7d4eab1243e",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Создание",
                "operationId": "1ac91d3e79872580c060ded21d68ec37",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lastname": {
                                        "type": "string",
                                        "example": "Тахиров"
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "example": "Далер"
                                    },
                                    "patronymic": {
                                        "type": "string",
                                        "example": "Рахимович"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Худжанд, 20мкр"
                                    },
                                    "phone": {
                                        "type": "integer",
                                        "example": 923002000
                                    },
                                    "salary": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "date_of_birthday": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-09-14"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Работник"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employees/{employee}": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Список по ID",
                "operationId": "8b29fc5dac0ba7532271df5cea20ac68",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "description": "ID клиент",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employee"
                ],
                "summary": "Удаление",
                "operationId": "4d9cce6625a76e79bf927c9684fab546",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "description": "ID клиент",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Employee"
                ],
                "summary": "Обновление",
                "operationId": "8761703579b0b3fb05bf4ff9d2939078",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "description": "ID клиент",
                        "required": true,
                        "example": 2
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lastname": {
                                        "type": "string",
                                        "example": "Тахиров"
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "example": "Далер"
                                    },
                                    "patronymic": {
                                        "type": "string",
                                        "example": "Рахимович"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Худжанд, 20мкр"
                                    },
                                    "phone": {
                                        "type": "integer",
                                        "example": 923002000
                                    },
                                    "salary": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "date_of_birthday": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-09-14"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Работник"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}